requirements.txt
pandas
sentence-transformers
google-generativeai
google-api-core
openai
python-dotenv
datasets
ragas
scikit-learn
pymupdf4llm
PyYAML